test(core): consolidate duplicate build-variant, app-registry, security, and service test suites (#28794)
